fixing bounding box issue + fixing build issue
