+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pdal-parallelizer
-version = 2.0.0
+version = 2.0.1
 description = A simple tool for use pdal with parallel execution
 long_description = file: README.rst, CHANGELOG.rst, LICENSE.rst
 keywords = pdal, parallelizer, dask
